fix(grottoRoutes): use v1 api + move official partners count route
organizations/official/count moves to partners/count url
