+++ api/controllers/PartnerController.js
@@ -43,6 +43,19 @@ module.exports = {
       });
   },
 
+  count: (req, res) => {
+    TGrotto.count({ isOfficialPartner: true }).exec((err, found) => {
+      const params = {
+        controllerMethod: 'PartnerController.count',
+        notFoundMessage: 'Problem while getting number of official partners.',
+      };
+      const count = {
+        count: found,
+      };
+      return ControllerService.treat(req, err, count, params, res);
+    });
+  },
+
   findForCarousel: (req, res) => {
     let skip = 0;
     if (req.param('skip')) {
