+++ api/controllers/v1/PartnerController.js
@@ -1,0 +1,15 @@
+/**
+ */
+
+const partnerController = require('../PartnerController');
+
+module.exports = {
+  findForCarousel: (req, res, next) =>
+    partnerController.findForCarousel(
+      req,
+      res,
+      next,
+      MappingV1Service.convertToOrganizationModel,
+    ),
+  count: (req, res) => partnerController.count(req, res),
+};
